Reduced amount of data visualized via charts
Due to performance issues related to the chart node of the dashboard, the trend charts are almost not used. The temperature chart is updated only when there is a difference in the new and alst measured temperature. The euro charts are updated only when there is new cost (i.s.o every time a new sample of E,W, or G is received).